aider: Fixed logo image not displaying by updating the src attribute to a relative path in the index.html file.
# Aider chat conversation:

USER: I want you to help me about my frontend part application, for some reason logo is not showing.

ASSISTANT: To assist you with the issue of the logo not showing in your frontend application, I'll need to see the relevant code. This could involve HTML, CSS, JavaScript, or even server configuration files, depending on how your frontend is set up.

Please add the file(s) that you suspect might be causing the issue with the logo to the chat, such as your HTML file where the logo is supposed to appear, any CSS files that style the logo, or JavaScript files that might manipulate it. Once I have access to the code, I'll be able to propose the necessary changes.

USER: I added these *read-write* files: /home/joao/ufpb/ControlVision/frontend/index.html, /home/joao/ufpb/ControlVision/frontend/assets/login/icone_senha-removebg-preview.png

ASSISTANT: Given the information you've provided, it seems that the logo image is not displaying correctly on your webpage. The issue might be related to the path to the image file in the `src` attribute of the `img` tag.

In the provided HTML file, the `src` attribute of the `img` tag is set to a local file system path:

```html
<img id="logo" src="/home/joao/ufpb/ControlVision/frontend/assets/login/icone_senha-removebg-preview.png" alt="Logo">
```

Web browsers cannot access local file system paths for security reasons. Instead, you should use a relative path or a URL that the web server can serve. If the image is located in the `assets/login` directory relative to the `index.html` file, you should change the `src` attribute to the relative path like so:

```diff
--- frontend/index.html
+++ frontend/index.html
@@ ... @@
-        <img id="logo" src="/home/joao/ufpb/ControlVision/frontend/assets/login/icone_senha-removebg-preview.png" alt="Logo">
+        <img id="logo" src="assets/login/icone_senha-removebg-preview.png" alt="Logo">
@@ ... @@
```

This change assumes that the `assets` directory is in the same directory as your `index.html` file. If the structure is different, you'll need to adjust the relative path accordingly. Make sure that the web server is configured to serve the files from the `assets` directory.